backend/requirements.txt
fastapi
uvicorn
pyyaml
pydantic
requests
